requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-allauth==0.40.0
django-extensions==2.2.5
django-filter==2.2.0
django-import-export==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-csv==2.1.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.3
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
unicode==2.7
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
drf-yasg

